requirements.txt
scikit-learn==1.3.0
pandas==2.1.0
nltk==3.8.1
python-dotenv==1.0.0
pyyaml==6.0.1
email-validator==2.0.0.post2
jinja2==3.1.2
pytest==7.4.0
fastapi==0.68.0
sqlalchemy==1.4.23
pydantic==1.8.2
transformers==4.11.0
torch==1.9.0
prometheus-client==0.11.0
uvicorn==0.15.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.5 